+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] / 2024-11-11 - 2024-11-19
+### Features
+- Support `tiff` files for icon.
+### Updated
+- Update `GetBitmapFrame` to round `width` that is changed by `dpi`.
+- Add `GetBitmapFrameByWidthAndDpi` to get optimal frame by `dpi` and `width`.
+- Add `GetSystemDpi` to get the system `dpi` on the fly.
+- Add `SystemDpi` to store the system `dpi` value.
+- Update `GetBitmapFrameByWidthAndDpi` to round `dpi` frame value.
+- Update `Width` to `Math.Round` to improve order by `Width`.
+- Update `GetBitmapFrameByWidthAndDpi` to public.
+- Update `GetBitmapFrameByWidthAndDpi` to return last `Width`.
+- Update `UriToBitmapFrame` to use `GetBitmapFrameByWidthAndDpi` with `int.MaxValue`.
+### Example
+- Add `Cube-Grey-Light.tiff` and `Cube-Grey-Dark.tiff` in `AppTheme`.
+### Tests
+- Add `ResourceTiffTests` with `Cube.tiff` with multiple dpi and scales.
+
 ## [0.7.0] / 2024-07-06 - 2024-07-25
 ### Features
 - Auto set image based on the theme of the Ribbon using `light` and `dark` image pattern.
@@ -360,6 +378,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.7.1]: ../../compare/0.7.0...0.7.1
 [0.7.0]: ../../compare/0.6.2...0.7.0
 [0.6.2]: ../../compare/0.6.1...0.6.2
 [0.6.1]: ../../compare/0.6.0...0.6.1